python/requirements.txt
cffi==1.17.0
ffmpeg-python==0.2.0
future==1.0.0
numpy==2.1.0
pycparser==2.22
pygame==2.6.0
sounddevice==0.5.0
